Add dependency on logic-model. Prepare game data for starting new game.

+++ logic-model/src/main/java/org/mishgan/color_capture/player/Player.java
@@ -6,8 +6,8 @@ public class Player {
     private byte colorNumber;
     private Point startPosition;
 
-    public Player colorNumber(byte colorNumber) {
-        this.colorNumber = colorNumber;
+    public Player colorNumber(int colorNumber) {
+        this.colorNumber = (byte) colorNumber;
         return this;
     }
